// src/app/types/Food/fooddata.ts
export type FoodData = {
    id: string;
    food_img: string;
    food_title: string;
    food_price: string;
    category: string;
    category_img: string;
    location: string;
    tag: string;
    status: string;
    type: string;
    rating: string;
    preparation_time: string;
    branch: string;
    slug: string;
    about?: string;
};